+++ src/main/java/superapp/logic/UsersServiceDB.java
@@ -1,7 +1,10 @@
 package superapp.logic;
 
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import superapp.boundaries.user.NewUserBoundary;
 import superapp.boundaries.user.UserBoundary;
@@ -10,17 +13,27 @@ import superapp.dal.UserCrud;
 import superapp.data.UserEntity;
 import superapp.data.UserRole;
 import superapp.utils.GeneralUtils;
-import superapp.utils.exceptions.AlreadyExistException;
-import superapp.utils.exceptions.InvalidInputException;
-import superapp.utils.exceptions.NotFoundException;
+import superapp.utils.exceptions.*;
+
 import java.util.*;
 import java.util.stream.Collectors;
 
+import static superapp.logic.ChatService.*;
+import static superapp.utils.Constants.DEFAULT_SORTING_DIRECTION;
+
+
+/**
+ * The UsersServiceDB class is an implementation of the UsersServiceAdvanced interface. It provides database-specific
+ * functionality for managing users.
+ * This class is annotated with the Spring framework's @Service annotation, indicating that it is a service component.
+ * It can be injected into other components or used as a dependency for managing user-related operations in the application.
+ */
 @Service
 public class UsersServiceDB implements UsersServiceAdvanced {
     private UserCrud userCrud; // CrudRepository
     private UserConverter converter; // Entity/Boundary converter
     private String superapp;
+    private final Log logger = LogFactory.getLog(UsersServiceDB.class);
 
     /**
      * Sets the ObjectCrud dependency.
@@ -42,6 +55,11 @@ public class UsersServiceDB implements UsersServiceAdvanced {
         this.converter = converter;
     }
 
+    /**
+     * Sets the value of the "spring.application.name" property to the "superapp" field.
+     *
+     * @param superapp The value of the "spring.application.name" property.
+     */
     @Value("${spring.application.name}")
     public void setSuperapp(String superapp) {
         this.superapp = superapp;
@@ -57,16 +75,26 @@ public class UsersServiceDB implements UsersServiceAdvanced {
      */
     @Override
     public UserBoundary createUser(UserBoundary user) {
+        this.logger.debug("-Creating new UserBoundary object");
         GeneralUtils.isValidNewUser(user.getUserId().getEmail(), user.getRole(), user.getUsername(), user.getAvatar());
 
         user.getUserId().setSuperapp(this.superapp);
-        if (this.userCrud.findById(this.converter.createID(user)).isPresent())
+        if (this.userCrud.findById(this.converter.createID(user.getUserId())).isPresent()) {
+            this.logger.error("-User {%s} Already Exist".formatted(user.getUserId()));
             throw new AlreadyExistException("User {%s} Already Exist".formatted(user.getUserId()));
-
+        }
         this.userCrud.save(this.converter.toEntity(user));
+        this.logger.trace("-Created New UserBoundary object ");
+
         return user;
     }
 
+    /**
+     * Creates a new user with the provided user data.
+     *
+     * @param user The user data for creating a new user.
+     * @return The created user.
+     */
     @Override
     public UserBoundary createUser(NewUserBoundary user) {
         return this.createUser(new UserBoundary(user.getEmail(), user.getRole(), user.getUsername(), user.getAvatar()));
@@ -83,14 +111,21 @@ public class UsersServiceDB implements UsersServiceAdvanced {
      */
     @Override
     public UserBoundary login(String userSuperApp, String userEmail) {
-        if (!userSuperApp.equals(this.superapp))
+        this.logger.debug("Trying to login");
+        if (!userSuperApp.equals(this.superapp)) {
+            this.logger.error("Invalid SuperApp, got {%s}, expected {%s}"
+                    .formatted(userSuperApp, this.superapp));
             throw new InvalidInputException("Invalid SuperApp, got {%s}, expected {%s}"
                     .formatted(userSuperApp, this.superapp));
+        }
 
         Optional<UserEntity> userCheck = this.userCrud.findById(this.converter.createID(userSuperApp, userEmail));
-        if (userCheck.isEmpty())
+        if (userCheck.isEmpty()) {
+            this.logger.error("User {%s, %s} Doesn't Exist".formatted(userSuperApp, userEmail));
             throw new NotFoundException("User {%s, %s} Doesn't Exist".formatted(userSuperApp, userEmail));
+        }
 
+        this.logger.trace("Login succeed");
         return this.converter.toBoundary(userCheck.get());
     }
 
@@ -106,9 +141,14 @@ public class UsersServiceDB implements UsersServiceAdvanced {
      */
     @Override
     public UserBoundary updateUser(String userSuperApp, String userEmail, UserBoundary update) {
+        this.logger.debug("Updating a UserBoundary object");
         Optional<UserEntity> userCheck = this.userCrud.findById(this.converter.createID(userSuperApp, userEmail));
-        if (userCheck.isEmpty()) throw new NotFoundException("User's Key {%s} doesn't exist"
-                .formatted(this.converter.createID(userSuperApp, userEmail)));
+        if (userCheck.isEmpty()) {
+            this.logger.error("User's Key {%s} doesn't exist"
+                    .formatted(this.converter.createID(userSuperApp, userEmail)));
+            throw new NotFoundException("User's Key {%s} doesn't exist"
+                    .formatted(this.converter.createID(userSuperApp, userEmail)));
+        }
 
         UserEntity user = userCheck.get();
         String newRole = update.getRole();
@@ -117,21 +157,54 @@ public class UsersServiceDB implements UsersServiceAdvanced {
 
         if (newRole != null) {
             if (GeneralUtils.isValidRole(newRole)) user.setRole(UserRole.valueOf(newRole));
-            else throw new InvalidInputException("Role {%s} Is Invalid".formatted(newRole));
+            else {
+                this.logger.error("Role {%s} Is Invalid".formatted(newRole));
+                throw new InvalidInputException("Role {%s} Is Invalid".formatted(newRole));
+            }
         }
 
         if (newUsername != null) {
             if (!newUsername.isEmpty()) user.setUsername(newUsername);
-            else throw new InvalidInputException("Username Can't Be Empty");
+            else {
+                this.logger.error("Username Can't Be Empty");
+                throw new InvalidInputException("Username Can't Be Empty");
+            }
         }
 
         if (newAvatar != null) {
             if (!newAvatar.isEmpty()) user.setAvatar(newAvatar);
-            else throw new InvalidInputException("Avatar Can't Be Empty");
+            else {
+                this.logger.error("Avatar Can't Be Empty");
+                throw new InvalidInputException("Avatar Can't Be Empty");
+            }
         }
 
         this.userCrud.save(user);
+        this.logger.trace("Updated a UserBoundary object");
         return this.converter.toBoundary(user);
+    }
+
+    /**
+     * Retrieves all users.
+     *
+     * @return a list of UserBoundary objects representing all the users
+     * @deprecated
+     */
+    @Override
+    @Deprecated
+    public List<UserBoundary> getAllUsers() {
+        this.logger.error("Method {getAllUsers()} is Deprecated");
+        throw new MethodNotInUseException("Method {getAllUsers()} is Deprecated");
+    }
+
+    /**
+     * @deprecated Deletes all users.
+     */
+    @Override
+    @Deprecated
+    public void deleteAllUsers() {
+        this.logger.error("Method {deleteAllUsers()} is Deprecated");
+        throw new MethodNotInUseException("Method {deleteAllUsers()} is Deprecated");
     }
 
     /**
@@ -140,16 +213,38 @@ public class UsersServiceDB implements UsersServiceAdvanced {
      * @return a List of all UserBoundary objects in the database
      */
     @Override
-    public List<UserBoundary> getAllUsers() {
-        return this.userCrud.findAll().stream().map(this.converter::toBoundary).collect(Collectors.toList());
+    public List<UserBoundary> getAllUsers(String userSuperApp, String userEmail, int size, int page) {
+        this.logger.debug("Getting all UserBoundary objects");
+        String userId = this.converter.createID(userSuperApp, userEmail);
+        PageRequest pageReq = PageRequest.of(page, size, DEFAULT_SORTING_DIRECTION, "id", "role");
+
+        if (GeneralUtils.isAuthUserOperation(userId, UserRole.ADMIN, userCrud)) {
+            this.logger.trace("Got all UserBoundary objects");
+            return this.userCrud
+                    .findAll(pageReq)
+                    .stream()
+                    .map(this.converter::toBoundary)
+                    .collect(Collectors.toList());
+        } else {
+            this.logger.error("Unauthorized User: Only Admin Can Get All Users");
+            throw new UnauthorizedUserOperation("Only Admin Can Get All Users");
+        }
     }
 
     /**
      * Deletes all user entities from the database.
      */
     @Override
-    public void deleteAllUsers() {
-        this.userCrud.deleteAll();
+    public void deleteAllUsers(String userSuperApp, String userEmail) {
+        this.logger.debug("Deleting all UserBoundary objects");
+        String userId = this.converter.createID(userSuperApp, userEmail);
+        if (GeneralUtils.isAuthUserOperation(userId, UserRole.ADMIN, userCrud)) {
+            this.userCrud.deleteAll();
+            this.logger.trace("Delete All Users Succeed");
+        } else {
+            this.logger.error("Unauthorized User: Only Admin Can Remove All Users");
+            throw new UnauthorizedUserOperation("Only Admin Can Remove All Users");
+        }
     }
 }
